FaceOrientationBackend/requirements.txt
fastapi
uvicorn
Pillow
numpy
opencv-python-headless
numpy
mediapipe
python-multipart